requirements.txt
uvicorn[standard]
fastapi
pydantic
python-dotenv
psycopg2-binary
python-multipart
face_recognition
opencv-python